Extract FormatOptionHandler from Display Builder into core-ui
