+++ app/display/model/src/main/java/org/csstudio/display/builder/model/Messages.java
@@ -45,16 +45,6 @@ public class Messages
     public static String FontStyle_BoldItalic;
     public static String FontStyle_Italic;
     public static String FontStyle_Regular;
-    public static String Format_Compact;
-    public static String Format_Decimal;
-    public static String Format_Default;
-    public static String Format_Engineering;
-    public static String Format_Exponential;
-    public static String Format_Hexadecimal;
-    public static String Format_Sexagesimal;
-    public static String Format_SexagesimalDMS;
-    public static String Format_SexagesimalHMS;
-    public static String Format_String;
     public static String GroupWidget_Description;
     public static String GroupWidget_Name;
     public static String Interpolation_Automatic;
