+++ app/display/model/src/main/java/org/csstudio/display/builder/model/widgets/TextEntryWidget.java
@@ -39,10 +39,10 @@ import org.csstudio.display.builder.model.persist.NamedWidgetFonts;
 import org.csstudio.display.builder.model.persist.WidgetColorService;
 import org.csstudio.display.builder.model.persist.WidgetFontService;
 import org.csstudio.display.builder.model.persist.XMLTags;
-import org.csstudio.display.builder.model.properties.FormatOption;
 import org.csstudio.display.builder.model.properties.WidgetColor;
 import org.csstudio.display.builder.model.properties.WidgetFont;
 import org.phoebus.framework.persistence.XMLUtil;
+import org.phoebus.ui.vtype.FormatOption;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
