+++ app/display/model/src/main/resources/org/csstudio/display/builder/model/messages.properties
@@ -29,16 +29,6 @@ FontStyle_Bold=Bold
 FontStyle_BoldItalic=Bold & Italic
 FontStyle_Italic=Italic
 FontStyle_Regular=Regular
-Format_Compact=Compact
-Format_Decimal=Decimal
-Format_Default=Default
-Format_Engineering=Engineering
-Format_Exponential=Exponential
-Format_Hexadecimal=Hexadecimal
-Format_Sexagesimal=Sexagesimal HH:MM:SS
-Format_SexagesimalDMS=Sexagesimal DMS 360deg rad
-Format_SexagesimalHMS=Sexagesimal HMS 24h rad
-Format_String=String
 GroupWidget_Description=Group of widgets
 GroupWidget_Name=Group
 Interpolation_Automatic=Automatic
