+++ core/ui/src/main/java/org/phoebus/ui/vtype/FormatOption.java
@@ -1,13 +1,13 @@
 /*******************************************************************************
- * Copyright (c) 2015-2016 Oak Ridge National Laboratory.
+ * Copyright (c) 2015-2018 Oak Ridge National Laboratory.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  *******************************************************************************/
-package org.csstudio.display.builder.model.properties;
+package org.phoebus.ui.vtype;
 
-import org.csstudio.display.builder.model.Messages;
+import org.phoebus.ui.Messages;
 
 /** Options for formatting a value
  *  @author Kay Kasemir
@@ -56,4 +56,5 @@ public enum FormatOption
     {
         return label;
     }
+
 }
