+++ core/ui/src/main/resources/org/phoebus/ui/messages.properties
@@ -2,6 +2,16 @@ AddColumn=Add Column
 AddRow=Add Row
 DefaultNewColumnName=New Column
 DoNotShow=Don't show this message again.
+Format_Compact=Compact
+Format_Decimal=Decimal
+Format_Default=Default
+Format_Engineering=Engineering
+Format_Exponential=Exponential
+Format_Hexadecimal=Hexadecimal
+Format_Sexagesimal=Sexagesimal HH:MM:SS
+Format_SexagesimalDMS=Sexagesimal DMS 360deg rad
+Format_SexagesimalHMS=Sexagesimal HMS 24h rad
+Format_String=String
 InstallExamples=Install Example Displays
 MagicLastRow=Click to add row
 MoveColumnLeft=Move Column Left
